requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
sqlmodel==0.0.16
python-jose==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
